[IMP] fieldservice_account: Simplify module
- This change removes most of the invoicing logic from fieldservice_account and puts it into its own module
