+++ fieldservice_account/readme/DESCRIPTION.rst
@@ -1,4 +1,3 @@
-This module adds the ability to track employee time and contractor
-costs for Field Service Orders. It also adds functionality to create
-a customer invoice and a vendor bill when a Field Service Order is
-completed.
+This module adds the ability to link field service orders to invoices.
+It also adds the option to track the billing partner for field service
+locations and orders.
