+++ fieldservice_account/views/account_invoice_view.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <record id="invoice_form_fsm_contractor" model="ir.ui.view">
