+++ fieldservice_account/views/fsm_location.xml
@@ -1,11 +1,10 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <!-- Field Service Location Tree View-->
     <record id="fsm_location_tree_view_accounting" model="ir.ui.view">
         <field name="name">fsm.location.tree.accounting</field>
         <field name="model">fsm.location</field>
-        <field name='inherit_id' ref='fieldservice.fsm_location_tree_view'/>
+        <field name="inherit_id" ref="fieldservice.fsm_location_tree_view"/>
         <field name="arch" type="xml">
             <field name="owner_id" position="after">
                 <field name="customer_id"/>
@@ -17,15 +16,10 @@
     <record id="fsm_location_form_view_accounting" model="ir.ui.view">
         <field name="name">fsm.location.form.accounting</field>
         <field name="model">fsm.location</field>
-        <field name='inherit_id' ref='fieldservice.fsm_location_form_view'/>
+        <field name="inherit_id" ref="fieldservice.fsm_location_form_view"/>
         <field name="arch" type="xml">
-            <page id="others-page" position='before'>
-                <page name='accounting' string='Accounting'>
-                    <group>
-                        <group>
-                            <field name='analytic_account_id'/>
-                        </group>
-                    </group>
+            <page id="others-page" position="before">
+                <page name="accounting" string="Accounting">
                 </page>
             </page>
             <field name="owner_id" position="after">
@@ -38,7 +32,7 @@
     <record id="fsm_location_search_view_accounting" model="ir.ui.view">
         <field name="name">fsm.location.search</field>
         <field name="model">fsm.location</field>
-        <field name='inherit_id' ref='fieldservice.fsm_location_search_view'/>
+        <field name="inherit_id" ref="fieldservice.fsm_location_search_view"/>
         <field name="arch" type="xml">
             <field name="owner_id" position="after">
                 <field name="customer_id"/>
