+++ fieldservice_account/views/fsm_person.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <record id="fsm_person_form" model="ir.ui.view">
@@ -6,16 +5,6 @@
         <field name="model">fsm.person</field>
         <field name="inherit_id" ref="fieldservice.fsm_person_form"/>
         <field name="arch" type="xml">
-            <div name="button_box" position="inside">
-                <button type="object"
-                        name="action_view_bills"
-                        class="oe_stat_button"
-                        icon="fa-credit-card"
-                        groups="base.group_user">
-                    <field name="bill_count" widget="statinfo"
-                           string="Vendor Bills"/>
-                </button>
-            </div>
             <field name="owner_id" position="after">
                 <field name="customer_id"/>
             </field>
