+++ fieldservice_account_analytic/__manifest__.py
@@ -2,30 +2,24 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': 'Field Service - Accounting',
-    'summary': 'Track invoices linked to Field Service orders',
+    'name': 'Field Service - Analytic Accounting',
+    'summary': """Track analytic accounts on Field Service locations
+                  and orders""",
     'version': '12.0.2.0.1',
     'category': 'Field Service',
     'author': 'Open Source Integrators, Odoo Community Association (OCA)',
     'website': 'https://github.com/OCA/field-service',
     'depends': [
-        'fieldservice',
-        'account',
+        'fieldservice_account',
+        'analytic',
+        'product',
     ],
     'data': [
+        'data/ir_rule.xml',
         'security/ir.model.access.csv',
-        'report/fsm_order_report_template.xml',
         'views/fsm_location.xml',
-        'views/fsm_order.xml',
-        'views/fsm_person.xml',
-        'views/fsm_route.xml',
-        'views/account_invoice_view.xml',
-    ],
-    'demo': [
-        'demo/fsm_location.xml',
     ],
     'license': 'AGPL-3',
     'development_status': 'Beta',
     'maintainers': ['osimallen'],
-    'pre_init_hook': 'pre_init_hook',
 }
