+++ fieldservice_account_analytic/models/fsm_location.py
@@ -1,0 +1,11 @@
+# Copyright (C) 2018 - TODAY, Open Source Integrators
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class FSMLocation(models.Model):
+    _inherit = 'fsm.location'
+
+    analytic_account_id = fields.Many2one('account.analytic.account',
+                                          string='Analytic Account')
