+++ fieldservice_account_analytic/readme/CONFIGURE.rst
@@ -1,0 +1,4 @@
+To configure this module, you need to:
+
+* Go to Field Service > Master Data > Locations
+* Create or select a location and set their analytic account
