+++ fieldservice_account_analytic/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Michael Allen <mallen@opensourceintegrators.com>
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+* Brian McMaster <brian@mcmpest.com>
