+++ fieldservice_account_analytic/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module adds the ability to track your analytic accounts on field
+service orders.
