+++ fieldservice_account_analytic/readme/INSTALL.rst
@@ -1,0 +1,1 @@
+No special installation instructions
