+++ fieldservice_isp_account/README.rst
@@ -1,0 +1,121 @@
+==========================
+Field Service - Accounting
+==========================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Ffield--service-lightgray.png?logo=github
+    :target: https://github.com/OCA/field-service/tree/12.0/fieldservice_account
+    :alt: OCA/field-service
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/field-service-12-0/field-service-12-0-fieldservice_account
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/264/12.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This module adds the ability to track employee time and contractor
+costs for Field Service Orders. It also adds functionality to create
+a customer invoice and a vendor bill when a Field Service Order is
+completed.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Installation
+============
+
+No special installation instructions
+
+Configuration
+=============
+
+To configure this module, you need to:
+
+* Go to Field Service > Master Data > Locations
+* Create or select a location and set their analytic account
+
+Usage
+=====
+
+To use the module:
+
+On a field service order, open the "Accounting" tab. Depending on 
+whether the logged in user or an employee or not, they will see 
+either a place to enter timesheet records or contractor costs.
+
+The total cost of the order is calculated based on the entries in 
+the employee timesheet entries and contractor costs.
+
+When an order is completed, a customer invoice will be generated for
+the employee time and the contractor costs. A vendor bill will be 
+created for the contractor costs.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/field-service/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/field-service/issues/new?body=module:%20fieldservice_account%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Open Source Integrators
+
+Contributors
+~~~~~~~~~~~~
+
+* Michael Allen <mallen@opensourceintegrators.com>
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+
+Other credits
+~~~~~~~~~~~~~
+
+The development of this module has been financially supported by:
+
+* Open Source Integrators <https://opensourceintegrators.com>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+.. |maintainer-osimallen| image:: https://github.com/osimallen.png?size=40px
+    :target: https://github.com/osimallen
+    :alt: osimallen
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-osimallen| 
+
+This module is part of the `OCA/field-service <https://github.com/OCA/field-service/tree/12.0/fieldservice_account>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
