+++ fieldservice_isp_account/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (C) 2018 Open Source Integrators
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import models
