+++ fieldservice_isp_account/views/account.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <!-- Show confirmed FSM Orders in Accounting -->
