+++ fieldservice_isp_account/views/hr_timesheet.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <record id="hr_timesheet.act_hr_timesheet_line" model="ir.actions.act_window">
